resolving all issues
